// ClientApp/src/types/generated-api.ts
//----------------------
// <auto-generated>
//     Generated using the NSwag toolchain v14.5.0.0 (NJsonSchema v11.4.0.0 (Newtonsoft.Json v13.0.0.0)) (http://NSwag.org)
// </auto-generated>
//----------------------

/* eslint-disable */
// ReSharper disable InconsistentNaming



export interface CompanyDto {
    id: string;
    name: string | undefined;
    addressId: string;
    vacancies?: VacancyDto[] | undefined;
}

export interface CreateVacancyDto {
    title: string | undefined;
    description: string | undefined;
    companyId: string;
}

export interface ProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;

    [key: string]: any;
}

export interface VacancyDto {
    id: string;
    title: string | undefined;
    description: string | undefined;
    companyId: string;
}